RM-8466 Change BocReferenceValue.PopulateBusinessObjectList() to load the value collection from the search service even when the DataSource is null.
IBusinessObjectReferenceProperty.SearchAvailableObjects(...) supports null values for the referencing business object. Given that the API can handle null values (and already did because DataSource.BusinessObject) could always have been null, there is no reason to not call the search service with a null value for the referencing business object when the BocReferenceValue.Property is set and only the BocReferenceValue.DataSource is not set.